create AMI to create instances, modular config

+++ ElasticHerd/ec2_herd.py
@@ -5,7 +5,8 @@ import configparser
 
 import node_setup
 
-def kill_nodes():
+
+def kill_nodes(ami_id_file="elasticHerdAMI.txt"):
     ec2 = boto3.resource('ec2')
     ec2_client = boto3.client('ec2')
 
@@ -21,20 +22,21 @@ def kill_nodes():
         Filters=[{'Name': 'instance-state-name', 'Values': ['running', 'stopped', 'pending']}])
     if instances is not None:
         for instance in instances:
-            for tag in instance.tags:
-                if tag["Key"] == "Name" and tag["Value"] == "elasticHerdInstance":
-                    print("Killing instance {}".format(instance.id))
-                    ec2_client.terminate_instances(InstanceIds=[instance.id])
+            if instance.tags is not None:
+                for tag in instance.tags:
+                    if tag["Key"] == "Name" and tag["Value"] == "elasticHerdInstance":
+                        print("Killing instance {}".format(instance.id))
+                        ec2_client.terminate_instances(InstanceIds=[instance.id])
 
-                    still_there = True
-                    while still_there:
-                        still_there = False
-                        instance = ec2.Instance(instance.id)
-                        if not instance.state["Name"] == "terminated":
-                            time.sleep(5)
+                        still_there = True
+                        while still_there:
+                            still_there = False
+                            instance = ec2.Instance(instance.id)
+                            if not instance.state["Name"] == "terminated":
+                                time.sleep(5)
 
-                    time.sleep(5)
-                    break
+                        time.sleep(5)
+                        break
 
     #
     # File system share
@@ -114,9 +116,18 @@ def kill_nodes():
                                         KeyName=k["KeyName"]
                                         )
 
+    #
+    # delete AMI
+    #
+    print("Deleting AMI...")
+    amis = ec2.images.filter(Filters=[{'Name': 'name', 'Values': ['elasticHerd_image']}]).all()
+
+    for ami in amis:
+        ami.deregister()
+        time.sleep(5)
 
 def create_ec2_nodes(node_count=3, config_file="./elastic_config.cfg"):
-    config = configparser.ConfigParser()
+    config = configparser.ConfigParser(allow_no_value=True, delimiters=["="])
     config.read(config_file)
     keyfile = config['elasticherd']['keyfile']
 
@@ -219,10 +230,73 @@ def create_ec2_nodes(node_count=3, config_file="./elastic_config.cfg"):
     #
 
     print("Creating instances...")
-    nodes = ec2.create_instances(
+
+    #
+    # First node
+    #
+    print("Creating and setting up first node")
+    node1 = ec2.create_instances(
                                     ImageId='ami-ed82e39e',
-                                    MinCount=node_count,
-                                    MaxCount=node_count,
+                                    MinCount=1,
+                                    MaxCount=1,
+                                    KeyName='elasticHerd',
+                                    SecurityGroupIds=[
+                                        internal_group.id,
+                                        external_group.id
+                                    ],
+                                    InstanceType='t2.micro',
+                                    Monitoring={
+                                        'Enabled': True
+                                    },
+                                    SubnetId=subnet["SubnetId"],
+                                )
+    ec2.create_tags(
+                        Resources=[node.id for node in node1],
+                        Tags=[
+                                {
+                                    'Key': 'Name',
+                                    'Value': 'elasticHerdInstance'
+                                },
+                             ]
+                      )
+
+    #
+    # Configure Instance
+    #
+    print("Configuring instance 1...")
+    for node in node1:
+        n = ec2.Instance(node.id)
+        while not n.state["Name"] == "running":
+            time.sleep(5)
+            n = ec2.Instance(node.id)
+
+        time.sleep(30)
+        print("Configuring first node, instance {}".format(node.id))
+        print("Public DNS of first node - {}".format(n.public_dns_name))
+        if "apt_packages" in config:
+            node_setup.install_apt_packages(n.public_dns_name, "ubuntu", keyfile, config["apt_packages"])
+        if "pip_packages" in config:
+            node_setup.install_pip_packages(n.public_dns_name, "ubuntu", keyfile, config["pip_packages"])
+        if "github_python_install" in config:
+            node_setup.install_github_packages(n.public_dns_name, "ubuntu", keyfile, config["github_python_install"])
+
+    image = node1[0].create_image(
+        Name='elasticHerd_image',
+        Description='Ubuntu based AMI for MPI Calculations',
+        NoReboot=False,
+    )
+
+    while not image.state == "available":
+        time.sleep(5)
+        image = ec2.Image(image.id)
+
+    #
+    # Creating all other nodes
+    #
+    nodes = ec2.create_instances(
+                                    ImageId=image.id,
+                                    MinCount=node_count-1,
+                                    MaxCount=node_count-1,
                                     KeyName='elasticHerd',
                                     SecurityGroupIds=[
                                         internal_group.id,
@@ -246,6 +320,13 @@ def create_ec2_nodes(node_count=3, config_file="./elastic_config.cfg"):
                              ]
                       )
 
+
+    #
+    # add first node to all others
+    #
+    for n in node1:
+        nodes.append(n)
+
     #
     # Configure Instances
     #
@@ -259,7 +340,6 @@ def create_ec2_nodes(node_count=3, config_file="./elastic_config.cfg"):
         time.sleep(30)
         print("Configuring instance {}".format(node.id))
         print("Public DNS - {}".format(n.public_dns_name))
-        node_setup.install_mpi_ecosystem_keyfile(n.public_dns_name, "ubuntu", keyfile)
         node_setup.mount_efs_share(n.public_dns_name, "ubuntu", keyfile,
                                    "$(curl -s http://169.254.169.254/latest/meta-data/placement/availability-zone).{}.efs.{}.amazonaws.com".format(efs_share["FileSystemId"],"eu-west-1"))
 
@@ -275,7 +355,7 @@ def create_ec2_nodes(node_count=3, config_file="./elastic_config.cfg"):
 
     node_setup.create_mpi_hosts_file(n.public_dns_name, "ubuntu", keyfile, [ec2.Instance(host.id).public_dns_name for host in nodes])
     node_setup.create_known_hosts_file(n.public_dns_name, "ubuntu", keyfile, [ec2.Instance(host.id).public_dns_name for host in nodes])
-    node_setup.pull_code_to_mpi_share(n.public_dns_name, "ubuntu", keyfile)
+    node_setup.pull_code_to_mpi_share(n.public_dns_name, "ubuntu", keyfile, config["github_clone"])
 
 
     for node in nodes[1:]:
@@ -298,12 +378,6 @@ def create_ec2_nodes(node_count=3, config_file="./elastic_config.cfg"):
         for node in nodes:
             f.write("{}\n".format(ec2.Instance(node.id).public_dns_name))
 
-    # instances = ec2.instances.filter(
-    #     Filters=[{'Name': 'instance-state-name', 'Values': ['running']}])
-    # for instance in instances:
-    #     print(instance.id, instance.instance_type)
-
-
 if __name__ == "__main__":
-    #kill_nodes()
-    create_ec2_nodes(node_count=10)
+    kill_nodes()
+    #create_ec2_nodes(node_count=3)
